자잘한 오류 수정
