+++ src/main/java/com/shop/farmers/base/security/CustomUserDetailsService.java
@@ -9,24 +9,24 @@ import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
-@Service
-@RequiredArgsConstructor
-public class CustomUserDetailsService implements UserDetailsService {
-    private final MemberRepository memberRepository;
-
-    //일반 로그인 시, 실행
-    @Override
-    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
-        Member member = memberRepository.findByEmail(email);
-
-        if(member == null){
-            throw new UsernameNotFoundException(email);
-        }
-
-        return User.builder()
-                .username(member.getEmail())
-                .password(member.getPassword())
-                .roles(member.getRole().toString())
-                .build();
-    }
-}
+//@Service
+//@RequiredArgsConstructor
+//public class CustomUserDetailsService implements UserDetailsService {
+//    private final MemberRepository memberRepository;
+//
+//    //일반 로그인 시, 실행
+//    @Override
+//    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
+//        Member member = memberRepository.findByEmail(email);
+//
+//        if(member == null){
+//            throw new UsernameNotFoundException(email);
+//        }
+//
+//        return User.builder()
+//                .username(member.getEmail())
+//                .password(member.getPassword())
+//                .roles(member.getRole().toString())
+//                .build();
+//    }
+//}
