+++ src/main/java/com/shop/farmers/boundedContext/home/controller/MainController.java
@@ -1,6 +1,5 @@
 package com.shop.farmers.boundedContext.home.controller;
 
-import com.shop.farmers.base.security.CustomUserDetailsService;
 import com.shop.farmers.boundedContext.item.dto.ItemSearchDto;
 import com.shop.farmers.boundedContext.item.dto.MainItemDto;
 import com.shop.farmers.boundedContext.item.service.ItemService;
@@ -24,7 +23,6 @@ import java.util.Optional;
 @Controller
 @RequiredArgsConstructor
 public class MainController {
-    private final CustomUserDetailsService customUserDetailsService;
     private final ItemService itemService;
     private final MemberService memberService;
 
@@ -48,7 +46,7 @@ public class MainController {
 
     // 권한 조회
     protected Authentication createNewAuthentication(Authentication currentAuth, String username) {
-        UserDetails newPrincipal = customUserDetailsService.loadUserByUsername(username);
+        UserDetails newPrincipal = memberService.loadUserByUsername(username);
         UsernamePasswordAuthenticationToken newAuth = new UsernamePasswordAuthenticationToken(newPrincipal, currentAuth.getCredentials(), newPrincipal.getAuthorities());
         newAuth.setDetails(currentAuth.getDetails());
         return newAuth;
