+++ src/main/java/com/shop/farmers/boundedContext/member/dto/MemberFormDto.java
@@ -11,7 +11,7 @@ import org.hibernate.validator.constraints.Length;
 @Setter
 public class MemberFormDto {
     @NotBlank(message = "이름은 필수 입력 값입니다.")
-    private String name;
+    private String username;
 
     @NotEmpty(message = "이메일은 필수 입력 값입니다.")
     @Email(message = "이메일 형식으로 입력해주세요.")
