+++ src/main/resources/templates/member/memberForm.html
@@ -30,9 +30,9 @@
 
     <form action="/members/new" role="form" method="post"  th:object="${memberFormDto}">
         <div class="form-group">
-            <label th:for="name">이름</label>
-            <input type="text" th:field="*{name}" class="form-control" placeholder="이름을 입력해주세요">
-            <p th:if="${#fields.hasErrors('name')}" th:errors="*{name}" class="fieldError">Incorrect data</p>
+            <label th:for="username">이름</label>
+            <input type="text" th:field="*{username}" class="form-control" placeholder="이름을 입력해주세요">
+            <p th:if="${#fields.hasErrors('username')}" th:errors="*{username}" class="fieldError">Incorrect data</p>
         </div>
         <div class="form-group">
             <label th:for="email">이메일주소</label>
